Ensure GeneratedFileReplacingCompilationTracker.ContainsAssemblyOrModuleOrDynamic works
This is a naive implementation that simply ensures we don't get
different behavior from the main implementation. Most of this change
is just moving the creation and matching logic into UnrootedSymbolSet.cs
itself.
